requirements.txt
pytest
selenium
webdriver-manager
requests
python-dotenv
allure-pytest